change doubles to floats

+++ basic_nn.cpp
@@ -3,22 +3,22 @@
 #include <algorithm>
 #include "basic_nn.hpp"
 
-neural_network::neural_network(int num_layers, std::vector<int>& layer_sizes, double learning_rate) {
+neural_network::neural_network(int num_layers, std::vector<int>& layer_sizes, float learning_rate) {
     assert(num_layers == layer_sizes.size());
     this -> num_layers = num_layers;
-    layer = (double **)malloc(sizeof(double *) * num_layers);
-    error = (double **)malloc(sizeof(double *) * num_layers);
+    layer = (float **)malloc(sizeof(float *) * num_layers);
+    error = (float **)malloc(sizeof(float *) * num_layers);
     for (int i = 0; i < num_layers; i++) {
-        layer[i] = (double *)calloc(layer_sizes[i], sizeof(double));
-        error[i] = (double *)calloc(layer_sizes[i], sizeof(double));
+        layer[i] = (float *)calloc(layer_sizes[i], sizeof(float));
+        error[i] = (float *)calloc(layer_sizes[i], sizeof(float));
     }
     // Xavier initialization
-    weights = (double ***)malloc(sizeof(double **) * (num_layers - 1));
+    weights = (float ***)malloc(sizeof(float **) * (num_layers - 1));
     for (int i = 1; i < num_layers; i++) {
-        weights[i - 1] = (double **)malloc(sizeof(double *) * layer_sizes[i - 1]);
-        double bounds = sqrt(6)/sqrt(layer_sizes[i - 1] + layer_sizes[i]);
+        weights[i - 1] = (float **)malloc(sizeof(float *) * layer_sizes[i - 1]);
+        float bounds = sqrt(6)/sqrt(layer_sizes[i - 1] + layer_sizes[i]);
         for (int j = 0; j < layer_sizes[i - 1]; j++) {
-            weights[i - 1][j] = (double *)malloc(sizeof(double) * layer_sizes[i]);
+            weights[i - 1][j] = (float *)malloc(sizeof(float) * layer_sizes[i]);
             for (int k = 0; k < layer_sizes[i]; k++) {
                 weights[i - 1][j][k] = randd(-bounds, bounds);
             }
@@ -28,7 +28,7 @@ neural_network::neural_network(int num_layers, std::vector<int>& layer_sizes, do
     this -> layer_sizes = layer_sizes;
 }
 
-void neural_network::forward_propagate(std::vector<double>& inputs) {
+void neural_network::forward_propagate(std::vector<float>& inputs) {
     assert(inputs.size() == layer_sizes[0]);
     for (int i = 0; i < inputs.size(); i++) {
         layer[0][i] = inputs[i];
@@ -45,7 +45,7 @@ void neural_network::forward_propagate(std::vector<double>& inputs) {
     }
 }
 
-void neural_network::backward_propagate(std::vector<double>& expected) {
+void neural_network::backward_propagate(std::vector<float>& expected) {
     assert(expected.size() == layer_sizes[num_layers - 1]);
     for (int i = 0; i < expected.size(); i++) {
         error[num_layers - 1][i] = expected[i] - layer[num_layers - 1][i];
@@ -63,12 +63,12 @@ void neural_network::backward_propagate(std::vector<double>& expected) {
     }
 }
 
-void neural_network::train(std::vector<double>& inputs, std::vector<double>& expected) {
+void neural_network::train(std::vector<float>& inputs, std::vector<float>& expected) {
     forward_propagate(inputs);
     backward_propagate(expected);
 }
 
-int neural_network::query(std::vector<double>& inputs) {
+int neural_network::query(std::vector<float>& inputs) {
     forward_propagate(inputs);
     return std::max_element(layer[num_layers - 1], layer[num_layers - 1] + layer_sizes[num_layers - 1]) - layer[num_layers - 1];
 }
