+++ basic_nn.hpp
@@ -6,19 +6,19 @@
 
 class neural_network {
 private:
-    double **layer;
-    double **error;
-    double ***weights;
+    float **layer;
+    float **error;
+    float ***weights;
     std::vector<int> layer_sizes;
-    double learning_rate;
+    float learning_rate;
     int num_layers;
-    void forward_propagate(std::vector<double>& inputs);
-    void backward_propagate(std::vector<double>& expected);
+    void forward_propagate(std::vector<float>& inputs);
+    void backward_propagate(std::vector<float>& expected);
 
 public:
-    neural_network(int num_layers, std::vector<int>& layer_sizes, double learning_rate);
-    void train(std::vector<double>& inputs, std::vector<double>& expected);
-    int query(std::vector<double>& inputs);
+    neural_network(int num_layers, std::vector<int>& layer_sizes, float learning_rate);
+    void train(std::vector<float>& inputs, std::vector<float>& expected);
+    int query(std::vector<float>& inputs);
 };
 
 #endif
